DRTII-420 Simplified mechanism for sending updated minutes to subscribers
