+++ server/src/main/scala/actors/minutes/QueueMinutesActor.scala
@@ -1,48 +1,39 @@
 package actors.minutes
 
-import actors.FlightsStateActor.HandleRecalculations
 import actors.SetDeploymentQueueActor
 import actors.minutes.MinutesActorLike.{MinutesLookup, MinutesUpdate}
-import akka.actor.{ActorRef, Cancellable}
+import actors.queues.QueueLikeActor.UpdatedMillis
+import akka.actor.ActorRef
 import drt.shared.CrunchApi.{CrunchMinute, DeskRecMinute, MinutesContainer}
 import drt.shared.Terminals.Terminal
 import drt.shared.{SDateLike, TQM}
-import services.RecalculationRequester
 
 import scala.concurrent.Future
-import scala.concurrent.duration._
 import scala.language.postfixOps
 
 class QueueMinutesActor(now: () => SDateLike,
                         terminals: Iterable[Terminal],
                         lookup: MinutesLookup[CrunchMinute, TQM],
                         lookupLegacy: MinutesLookup[CrunchMinute, TQM],
-                        updateMinutes: MinutesUpdate[CrunchMinute, TQM]) extends MinutesActorLike(now, terminals, lookup, lookupLegacy, updateMinutes) {
+                        updateMinutes: MinutesUpdate[CrunchMinute, TQM])
+  extends MinutesActorLike(now, terminals, lookup, lookupLegacy, updateMinutes) {
 
-  val reDeployHandler = new RecalculationRequester()
-
-  val cancellableTick: Cancellable = context.system.scheduler.schedule(10 seconds, 1 second, self, HandleRecalculations)
-
-  override def postStop(): Unit = {
-    log.warn("Actor stopped. Cancelling scheduled tick")
-    cancellableTick.cancel()
-    super.postStop()
-  }
+  var maybeUpdatesSubscriber: Option[ActorRef] = None
 
   override def handleUpdatesAndAck(container: MinutesContainer[CrunchMinute, TQM],
                                    replyTo: ActorRef): Future[Option[MinutesContainer[CrunchMinute, TQM]]] = {
     val eventualUpdatesDiff = super.handleUpdatesAndAck(container, replyTo)
     val gotDeskRecs = container.contains(classOf[DeskRecMinute])
 
-    if (gotDeskRecs) addUpdatesToBufferAndSendToSubscriber(eventualUpdatesDiff).onComplete(_ => self ! HandleRecalculations)
+    if (gotDeskRecs) sendUpdatedMillisToSubscriber(eventualUpdatesDiff)
 
     eventualUpdatesDiff
   }
 
-  private def addUpdatesToBufferAndSendToSubscriber(eventualUpdatesDiff: Future[Option[MinutesContainer[CrunchMinute, TQM]]]): Future[Unit] = eventualUpdatesDiff.collect {
+  private def sendUpdatedMillisToSubscriber(eventualUpdatesDiff: Future[Option[MinutesContainer[CrunchMinute, TQM]]]): Future[Unit] = eventualUpdatesDiff.collect {
     case Some(diffMinutesContainer) =>
       val updatedMillis = diffMinutesContainer.minutes.collect { case m: CrunchMinute => m.minute }
-      reDeployHandler.addMillis(updatedMillis.toSeq)
+      maybeUpdatesSubscriber.foreach(_ ! UpdatedMillis(updatedMillis))
   }
 
   override def receive: Receive = deploymentReceives orElse super.receive
@@ -50,8 +41,6 @@ class QueueMinutesActor(now: () => SDateLike,
   def deploymentReceives: Receive = {
     case SetDeploymentQueueActor(subscriber) =>
       log.info(s"Received subscriber actor")
-      reDeployHandler.setQueueActor(subscriber)
-
-    case HandleRecalculations => reDeployHandler.handleUpdatedMillis()
+      maybeUpdatesSubscriber = Option(subscriber)
   }
 }
