--- server/src/test/scala/services/RequestRecalculationSpec.scala
@@ -1,9 +1,0 @@
-package services
-
-import services.crunch.CrunchTestLike
-
-
-
-class RequestRecalculationSpec extends CrunchTestLike {
-
-}
